MOD: rename sample files and add comments
