+++ apis/infix-en_GB/atrium-api-infix-en_GB/src/commonTest/kotlin/ch/tutteli/atrium/api/infix/en_GB/samples/IterableSubjectChangerSamples.kt
@@ -8,10 +8,11 @@ import ch.tutteli.atrium.api.infix.en_GB.toEqual
 import ch.tutteli.atrium.api.verbs.internal.expect
 import kotlin.test.Test
 
-class IterableSubjectChangersSamples {
+class IterableSubjectChangerSamples {
     @Test
     fun asListFeature() {
         expect(0..2) asList o toEqual listOf(0, 1, 2)
+        //                           | subject is now of type List<Int>
 
         fails {
             expect(0..2) asList o toContain 3 toContain 4
@@ -21,15 +22,18 @@ class IterableSubjectChangersSamples {
     @Test
     fun asList() {
         expect(0..2)
-            .asList {
+            .asList { // subject within this expectation-group is of type List<Int>
                 it toEqual listOf(0, 1, 2)
-            }
+            }  // subject here is back to type IntRange
 
         fails {
+            // all expectations inside an expectation-group are evaluated together; for more details see:
+            // https://github.com/robstoll/atrium#define-single-expectations-or-an-expectation-group
             expect(0..2)
                 .asList {
-                    it toContain 3
-                    it toContain 4
+                    it toContain 3  // fails
+                    it toContain 4  // still evaluated even though above `toContain` already fails
+                    //                   use `asList o` if you want a fail fast behaviour
                 }
         }
     }
